Implementing Censor
Replacing previous "hidden_fields", censor allow the definition of "private" and "public"
field for each model. For each model, only the field defined in public, or not
defined in private, will be present in the resulting rows. If both private
and public field are defined for a same model, only the public definition will
be used.
One can also decide to use django's permission implementation. Dgeq will check for
the view_[model] permission for each related field. Note that permission can be used
in conjunction with private and public field.

Finally, if the user manually try to access a private field (or is missing a permission),
the UnknownFieldError will be raised.